@@ -1,16 +1,16 @@
 import 'package:flutter/material.dart';
 
 class MenuItem extends StatelessWidget {
-  final IconData icon;
-  final String title;
-  final Function onTap;
+  final IconData? icon;
+  final String? title;
+  final Function? onTap;
 
-  const MenuItem({Key key, this.icon, this.title, this.onTap}) : super(key: key);
+  const MenuItem({Key? key, this.icon, this.title, this.onTap}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
     return GestureDetector(
-      onTap: onTap,
+      onTap: onTap as void Function()?,
       child: Padding(
         padding: const EdgeInsets.all(16),
         child: Row(
@@ -24,7 +24,7 @@ class MenuItem extends StatelessWidget {
               width: 20,
             ),
             Text(
-              title,
+              title!,
               style: TextStyle(fontWeight: FontWeight.w300, fontSize: 26, color: Colors.white),
             )
           ],
